PT-2498: Make pedigree editor extensible Fixed
